feat: runQuery 함수 생성
